Added new tests
Fixed missing @Valid annotation, replaced @NotNull annotation with @NotBlank annotation

+++ src/main/java/com/pet/petmanagement/domain/pet/PetRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface PetRepository extends JpaRepository<Pet, Long> {
+public interface PetRepository extends JpaRepository<Pet, Integer> {
     @Query("select p from Pet p where p.user.id = ?1")
     List<Pet> findAllUserPetsByUserId(Integer id);
