+++ src/main/java/com/pet/petmanagement/domain/pet/PetService.java
@@ -15,7 +15,7 @@ public class PetService {
         petRepository.save(pet);
     }
 
-    public Pet findById(Long petId) {
+    public Pet findById(Integer petId) {
        return petRepository.findById(petId).get();
     }
